Brain-Tumor-Detection---MRI-main/requirements.txt
tensorflow==2.13.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
pillow==10.0.1
scikit-learn==1.3.0
gdown==4.7.1
streamlit==1.32.0